requirements.txt
django
djangorestframework
psycopg2
python-dotenv
requests